Combine common Phone, Email and Address class functions into a superclass Contact and make them children
